requirements.txt
pystray>=0.17.3
pillow>=8.2.0
psutil>=5.8.0
